+++ jest.config.js
@@ -1,0 +1,17 @@
+const path = require('path')
+
+module.exports = {
+  verbose: true,
+  transform: {
+    '^.+\\.tsx?$': 'ts-jest',
+  },
+  setupFiles: [path.join(__dirname, './testing/setup-tests.js')],
+  coverageThreshold: {
+    global: {
+      branches: 100,
+      functions: 100,
+      lines: 100,
+      statements: 100,
+    },
+  },
+}