+++ plugins/graphql/src/index.ts
@@ -1,3 +1,0 @@
-export const graphql = () => {
-  // To be implemented
-}