+++ plugins/json-schema/src/index.spec.ts
@@ -1,7 +1,0 @@
-import { jsonSchema } from '.'
-
-describe('jsonSchema', () => {
-  test('test goes here', () => {
-    // To be implemented
-  })
-})