+++ plugins/json-schema/src/index.ts
@@ -1,3 +1,0 @@
-export const jsonSchema = () => {
-  // To be implemented
-}