+++ plugins/record-hook/src/index.ts
@@ -1,24 +1,4 @@
-import { Client, FlatfileEvent } from '@flatfile/listener'
-import { RecordHook } from '@flatfile/configure'
-import type { FlatfileRecord } from '@flatfile/hooks'
+export * from './RecordHook'
+export * from './record.hook.plugin'
 
-export const recordHook = (
-  sheetSlug: string,
-  callback: (record: FlatfileRecord) => {}
-) => {
-  return (client: Client) => {
-    client.on(
-      'records:*',
-      {
-        // todo: fix this filter
-        context: {
-          // @ts-ignore
-          sheetSlug,
-        },
-      },
-      (event: FlatfileEvent) => {
-        return RecordHook(event, callback)
-      }
-    )
-  }
-}
+export { FlatfileRecord } from '@flatfile/hooks'