+++ plugins/tsv-extractor/src/index.ts
@@ -1,0 +1,10 @@
+import { Flatfile } from '@flatfile/api'
+import { DelimiterExtractor } from '@flatfile/plugin-delimiter-extractor'
+
+export const TSVExtractor = (options?: {
+  dynamicTyping?: boolean
+  skipEmptyLines?: boolean | 'greedy'
+  transform?: (value: any) => Flatfile.CellValueUnion
+}) => {
+  return DelimiterExtractor('.tsv', { delimiter: '\t', ...options })
+}