+++ plugins/dxp-configure/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @flatfile/plugin-dxp-config
 
+## 0.0.5
+
+### Patch Changes
+
+- 7fa7925: updating readmes
+
+## 0.0.4
+
+### Patch Changes
+
+- 41261d0: Build tooling refactor
+
 ## 0.0.3
 
 ### Patch Changes